[Horus] S.PORT devices update should work (not tested) [Horus] SD Manager files renaming fixed (quick & dirty, I would prefer a popup)
