+++ radio/src/telemetry/frsky.cpp
@@ -165,7 +165,7 @@ NOINLINE void processFrskyTelemetryData(uint8_t data)
 
 #if defined(TELEMETRY_FRSKY_SPORT)
   if (IS_FRSKY_SPORT_PROTOCOL() && telemetryRxBufferCount >= FRSKY_SPORT_PACKET_SIZE) {
-    processSportPacket(telemetryRxBuffer);
+    sportProcessPacket(telemetryRxBuffer);
     dataState = STATE_DATA_IDLE;
   }
 #endif
